pyproject.toml
[project]
name = "pt-backend-fastapi"
version = "0.1.0"
description = "Prueba Tecnica Backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.120.1",
    "alembic>=1.12.1",
    "sqlalchemy[asyncio]>=2.0.23",
    "asyncpg>=0.29.0",
    "email-validator>=2.1.0.post1",
    "pydantic[email]>=2.4.2",
    "pydantic-settings>=2.11.0",
    "typing>=3.10.0.0",
    "pyjwt>=2.10.1",
    "pwdlib[argon2]>=0.3.0",
    "python-jose[cryptography]>=3.5.0",
]
